rename `lift`
